Update catchup branch with latest changes

+++ Hexagon/View_Models/SwipeableTaskDetailViewModel.swift
@@ -7,26 +7,40 @@
 
 import SwiftUI
 import HexagonData
+import Combine
 
 @MainActor
 class SwipeableTaskDetailViewModel: ObservableObject {
     @Published var reminders: [Reminder]
     @Published var lastUpdatedIndex: Int?
+
+    private var cancellables = Set<AnyCancellable>()
     
     init(reminders: [Reminder]) {
         self.reminders = reminders
+
+        $reminders
+            .sink { [weak self] _ in
+                self?.objectWillChange.send()
+            }
+            .store(in: &cancellables)
     }
     
+    // MARK: - Async/Await functions
+
     func deleteReminder(at index: Int, reminderService: ReminderService) async throws {
+        guard index < reminders.count else { return }
         let reminderToDelete = reminders[index]
+        
         try await reminderService.deleteReminder(reminderToDelete)
-        self.reminders.remove(at: index)
+        reminders.remove(at: index)
+        objectWillChange.send()
     }
-    
+
     func updateReminder(at index: Int, with updatedReminder: Reminder, tags: [String], photos: [UIImage]) {
         guard index < reminders.count else { return }
         reminders[index] = updatedReminder
         lastUpdatedIndex = index
-        objectWillChange.send()
+        objectWillChange.send() 
     }
 }
